Removed extra versions of Flowable, leaving just one
Deleted examples and tests that are covered in existing tests; also
commented out a few, which will need to be redone.  Those are still
present in the code, as a reminder of what needs to be ported.
