+++ examples/conditional-request-handling/ConditionalRequestHandling_Client.cpp
@@ -6,8 +6,8 @@
 #include "examples/util/ExampleSubscriber.h"
 #include "rsocket/RSocket.h"
 #include "rsocket/transports/TcpConnectionFactory.h"
-#include "yarpl/v/Flowable.h"
-#include "yarpl/v/Subscriber.h"
+#include "yarpl/Flowable.h"
+#include "yarpl/Subscriber.h"
 
 using namespace ::reactivesocket;
 using namespace ::folly;
