+++ examples/stream-hello-world/StreamHelloWorld_Client.cpp
@@ -9,8 +9,8 @@
 #include "rsocket/RSocket.h"
 #include "rsocket/transports/TcpConnectionFactory.h"
 
-#include "yarpl/v/Flowable.h"
-#include "yarpl/v/Subscriber.h"
+#include "yarpl/Flowable.h"
+#include "yarpl/Subscriber.h"
 
 using namespace reactivesocket;
 using namespace rsocket_example;
