+++ experimental/rsocket-test/handlers/HelloStreamRequestHandler.cpp
@@ -2,7 +2,7 @@
 
 #include "HelloStreamRequestHandler.h"
 #include <string>
-#include "yarpl/v/Flowables.h"
+#include "yarpl/Flowables.h"
 
 using namespace ::reactivesocket;
 using namespace yarpl;
