+++ experimental/rsocket/RSocketRequester.h
@@ -4,8 +4,7 @@
 
 #include <folly/io/async/EventBase.h>
 
-// TODO migrate yarpl/v/Flowable to yarpl/Flowable
-#include "yarpl/v/Flowable.h"
+#include "yarpl/Flowable.h"
 
 #include "src/ReactiveStreamsCompat.h"
 #include "src/ReactiveSocket.h"
