+++ experimental/yarpl/CMakeLists.txt
@@ -34,58 +34,33 @@ add_library(
         # public API
         include/yarpl/Scheduler.h
         include/yarpl/Disposable.h
+        # Flowable public API
+        include/yarpl/Flowable.h
+        include/yarpl/Flowables.h
+        include/yarpl/Operator.h
+        include/yarpl/Refcounted.h
+        include/yarpl/Subscriber.h
+        include/yarpl/Subscribers.h
+        include/yarpl/Subscription.h
         # Observable public API
         include/yarpl/Observable.h
         include/yarpl/Observable_Observer.h
         include/yarpl/Observable_Subscription.h
         include/yarpl/Observable_TestObserver.h
         src/yarpl/observable/Observable_Subscription.cpp
+        # Flowable sources
+        src/yarpl/Refcounted.cpp
         # Observable sources
         src/yarpl/observable/sources/Observable_RangeSubscription.h
         src/yarpl/observable/sources/Observable_RangeSubscription.cpp
         # Observable operators
         src/yarpl/observable/operators/Observable_Map.h
         src/yarpl/observable/operators/Observable_Take.h
-        # Flowable public API
-        include/yarpl/Flowable.h
-        include/yarpl/FlowableA.h
-        include/yarpl/FlowableB.h
-        include/yarpl/FlowableC.h
-        include/yarpl/Flowable_Subscriber.h
-        include/yarpl/Flowable_SubscriptionSync.h
-        include/yarpl/Flowable_TestSubscriber.h
-        # hack-y extraction of test code
-        include/yarpl/Tuple.h
-        # Flowable sources
-        src/yarpl/flowable/sources/Flowable_RangeSubscription.h
-        src/yarpl/flowable/sources/Flowable_RangeSubscription.cpp
-        src/yarpl/flowable/sources/Flowable_FromObservable.h
-        # Flowable operators
-        src/yarpl/flowable/operators/Flowable_Map.h
-        src/yarpl/flowable/operators/Flowable_Take.h
-        src/yarpl/flowable/operators/Flowable_SubscribeOn.h
-        # Flowable utils
-        src/yarpl/flowable/utils/SubscriptionHelper.h
-        src/yarpl/flowable/utils/SubscriptionHelper.cpp
         # utils
         src/yarpl/utils/type_traits.h
         # Scheduler
         src/yarpl/ThreadScheduler.cpp
         src/yarpl/ThreadScheduler.h
-        # move this out at some point
-        include/reactivestreams/ReactiveStreams.h
-        # hack-y extraction of test code
-        src/yarpl/Tuple.cpp
-        # The "v" version of Flowable.
-        include/yarpl/v/Flowable.h
-        include/yarpl/v/Flowables.h
-        include/yarpl/v/FlowableV.h
-        include/yarpl/v/Operator.h
-        include/yarpl/v/Refcounted.h
-        include/yarpl/v/Subscriber.h
-        include/yarpl/v/Subscribers.h
-        include/yarpl/v/Subscription.h
-        src/yarpl/v/Refcounted.cpp
 )
 
 target_include_directories(
@@ -98,17 +73,9 @@ target_include_directories(
 add_executable(
         yarpl-playground
         examples/yarpl-playground.cpp
-        examples/ObservableExamples.cpp
+#        examples/ObservableExamples.cpp
         examples/FlowableExamples.cpp
         examples/FlowableExamples.h
-        examples/FlowableAExamples.cpp
-        examples/FlowableAExamples.h
-        examples/FlowableBExamples.cpp
-        examples/FlowableBExamples.h
-        examples/FlowableCExamples.cpp
-        examples/FlowableCExamples.h
-        examples/FlowableVExamples.cpp
-        examples/FlowableVExamples.h
 )
 
 target_link_libraries(
@@ -124,21 +91,17 @@ target_include_directories(
 add_executable(
         yarpl-tests
         test/yarpl-tests.cpp
-        test/Observable_test.cpp
-        test/FlowableCreateSubscribe_test.cpp
-        test/Observable_Subscription_test.cpp
-        test/Flowable_range_test.cpp
-        test/FlowableSubscriptionHelper_test.cpp
-        test/Flowable_lifecycle.cpp
-        test/FlowableA_lifecycle.cpp
-        test/FlowableB_lifecycle.cpp
-        test/FlowableChaining_test.cpp
-        test/flowable_operators/Flowable_Map_test.cpp
-        test/flowable_operators/Flowable_Take_test.cpp
-        test/flowable_operators/Flowable_SubscribeOn_test.cpp
+#        test/Observable_test.cpp
+#        test/FlowableCreateSubscribe_test.cpp
+#        test/Observable_Subscription_test.cpp
+#        test/Flowable_range_test.cpp
+#        test/Flowable_lifecycle.cpp
+#        test/FlowableA_lifecycle.cpp
+#        test/FlowableB_lifecycle.cpp
+#        test/FlowableChaining_test.cpp
         test/Scheduler_test.cpp
-        test/v/RefcountedTest.cpp
-        test/v/FlowableTest.cpp
+        test/RefcountedTest.cpp
+        test/FlowableTest.cpp
 )
 
 target_link_libraries(
