--- experimental/yarpl/examples/FlowableBExamples.h
@@ -1,12 +1,0 @@
-// Copyright 2004-present Facebook. All Rights Reserved.
-
-#pragma once
-
-#include <iostream>
-#include <memory>
-#include <string>
-
-class FlowableBExamples {
- public:
-  static void run();
-};
