+++ experimental/yarpl/examples/yarpl-playground.cpp
@@ -1,26 +1,13 @@
 // Copyright 2004-present Facebook. All Rights Reserved.
 
 #include <iostream>
-#include "FlowableAExamples.h"
-#include "FlowableBExamples.h"
-#include "FlowableCExamples.h"
+
 #include "FlowableExamples.h"
-#include "FlowableVExamples.h"
-#include "ObservableExamples.h"
 
 int main() {
   std::cout << "*** Run yarpl::flowable::v examples ***" << std::endl;
-  // TODO: enable
-  // FlowableVExamples::run();
+  FlowableExamples::run();
   
   //  std::cout << "*** Run ObservableExamples ***" << std::endl;
   //  ObservableExamples::run();
-  //  std::cout << "*** Run FlowableExamples ***" << std::endl;
-  //  FlowableExamples::run();
-  //  std::cout << "*** Run FlowableExamplesA ***" << std::endl;
-  //  FlowableAExamples::run();
-  //  std::cout << "*** Run FlowableExamplesB ***" << std::endl;
-  //  FlowableBExamples::run();
-  //  std::cout << "*** Run FlowableExamplesC ***" << std::endl;
-  //  FlowableCExamples::run();
 }
